issue #312
git-svn-id: http://tradelink.googlecode.com/svn/trunk@1341 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ TradeLinkTests/TestTikWriterReader.cs
@@ -160,15 +160,17 @@ namespace TestTradeLink
             readdata.Add(t);
         }
 
-        static void removefile(string FILE)
+        public static bool removefile(string FILE)
         {
+            if (!System.IO.File.Exists(FILE)) return true;
             try
             {
                 System.IO.File.Delete(FILE);
+                return true;
             }
             catch (Exception ex) 
-            { 
-                System.Diagnostics.Debug.WriteLine(ex.Message); 
+            {
+                return false;
             }
         }
